fix: 消除 Codacy 静态分析告警
- 涉事文件改用相对路径导入：Codacy 引擎不解析 @ 别名，导入值被标记为
  error 类型并误报 no-unsafe-*（同分支使用 ./ 导入的 user.ts 无任何告警）
- showToast 从 ToastComponent.vue 抽离到 utils/toast.ts，逻辑与视图分离，
  组件内保留再导出以兼容既有调用方
- main.ts / SessionManagerComponent 补全 Promise 处理（void 标记 / await），
  refresh 的状态赋值移入 try 块
- api.ts 鉴权回调改为无参类型，消除未使用的 error 形参告警
- cookie.ts 用 String#startsWith 替代 indexOf === 0
